pyproject.toml
[project]
name = "lasalle-catalog"
version = "0.1.0"
description = "Local mirror and wiki source for the La Salle Campus Barcelona academic catalog"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "lxml>=5.0",
    "tenacity>=8.2",
    "typer>=0.12",
    "markdownify>=0.13",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts", "catalog_wiki_api"]
